[S5P6818/Build] change build in docker

+++ build.sh
@@ -6,8 +6,17 @@
 #########################################################################
 #!/bin/bash
 make s5p6818_gec6818_v1_defconfig
+if [ $? != 0 ];then
+	exit 1
+fi
 make CROSS_COMPILE=aarch64-linux-gnu-
-mv fip-nonsecure.img ./target/
+if [ $? != 0 ];then
+	exit 1
+fi
+mv -f fip-nonsecure.img ./target/
+if [ $? != 0 ];then
+	exit 1
+fi
 cd target
 dd if=bl1-mmcboot.bin of=gec6818_u-boot_with_fip.img bs=512 seek=0 conv=fdatasync
 dd if=fip-loader.img of=gec6818_u-boot_with_fip.img bs=512 seek=128 conv=fdatasync
@@ -15,5 +24,5 @@ dd if=fip-secure.img of=gec6818_u-boot_with_fip.img bs=512 seek=768 conv=fdatasy
 dd if=fip-nonsecure.img of=gec6818_u-boot_with_fip.img bs=512 seek=3840 conv=fdatasync
 sync
 mkdir -p output
-mv gec6818_u-boot_with_fip.img ./output/
+mv -f gec6818_u-boot_with_fip.img ./output/
 cd -
